Abstract import logic into base import module
