+++ medical_pathology_import/__init__.py
@@ -2,4 +2,4 @@
 # Copyright 2016 LasLabs Inc.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import medical_pathology_import
+from . import wizards
